Complete timer when past due.
It has been reported that when a browser tab is inactive or when a computer
is put to sleep, and it's long enough that the timer is past due, then it
would start counting back from 24:00:00.

This commit fixes that, causing the timer to immediatelly
complete when it's past due.

fix #16